clean up extension schema and docs (#9569)
* fix extension searchFilters item type

The JSON Schema type for the searchFilters contribution array item should be an object (because it has `name` and `value` properties), but it was incorrectly typed as an array. This type of contribution is rare, which is probably why this bug was not noticed earlier. It would have caused a validation warning when editing the extension's package.json file (in some cases), but otherwise would have had no other negative consequences.

* clean up extension schema documentation

- Removes unnecessary newlines from descriptions. These were introduced when the JSON Schema was auto-generated from TypeScript interfaces (whose tsdocs had newlines).
- Removes excessive duplication with https://docs.sourcegraph.com/extensions/authoring/contributions.
- Removes client directives (i.e., documentation about these fields that is intended for extension clients, such as the Sourcegraph web app and our code host integrations). That documentation belongs in the TypeScript implementation code, not in this JSON Schema that is intended for extension authors.