src-expose: test http handler (#9589)
* src-expose: test http handler

This adds tests for "src-expose serve", which currently has no test
coverage. This is in preperation of a slightly more tricky change.

* src-expose: write post-update hook

Rather than renaming the sample, lets just write the hook we want. This is
motivated by CI not having the sample hook.

* pass in wanted Repos

this makes it possible to change the URI we expect. This will be useful for
the follow-up PR which adds functionality. Keeping it out of that PR to make
things cleaner.