ci: run slowest steps first (#9614)
This has been updated with recent timings. Times have changed and new steps
have been added. The timings are unscientifically collected by looking at a
random build. Steps are assigned agents in the order they are specified. So as
a performance improvement we want to greedily run the slower steps first.