observability: more clear grouping support (#9642)
* observability: improved group support

Previously there were just "Rows" and if they had the same Title ("General") they would
implicitly become a part of the same group. This was hacky / smelly. Instead, this PR
changes that to make groups explicitly defined which will make dashboards that require
multiple groups easier to reason about (such as the frontend, which has "Search at a glance"
and "Internal service connections" groups, with multiple rows).

* observability: add support for hiding groups by default

* observability: ensure dashboards have changing versions for proper updating

If the version has not changed, dynamic updates will sometimes fail. This
only happens in development mode -- but is also a good idea to have in
production builds anyway in case Grafana's behavior in the future changes.