doc: update add repo from local disk instructions
Note how different the instructions are here. We require the user to configure
the code host first, then shutdown Sourcegraph. This is to prevent a
background Sourcegraph job removing the clone. If the code cost config doesn't
exist a background job could delete the repo thinking it shouldn't be
there. If you keep Sourcegraph running we may delete the repo since it will
appear corrupt while cloning.

We may want to invest in improving this experience. For example we could add
support for a "staging" area on gitserver. This can be consulted for clones.