observability: convert Repo Updater dashboard to generated
This PR is stacked on top of https://github.com/sourcegraph/sourcegraph/issues/9662 -- please review that first.

This is a direct translation of the current Repo Updater dashboard and rules to the new generator.

Before:

![image](https://user-images.githubusercontent.com/3173176/78826945-3128c700-7997-11ea-9674-d02b7e28cb7b.png)

After:

![image](https://user-images.githubusercontent.com/3173176/78826987-3e45b600-7997-11ea-9d4f-0f3c26a14559.png)

Note: the "Before" was taken on Sourcegraph.com which is why there is no datapoint there.