server: use /etc/sourcegraph/gitconfig
This follows the same pattern we have for SSH and netrc. It makes it much more
convenient for administrators to specify a git config for the Sourcegraph
Docker image.

Note: Cluster based deploys don't support this method, but those have ways to
put the configuration file in the correct place via resource definitions.