tracking-issue: Track Pull requests
This changeset adds pull request support to the tracking-issue tool. It
does a lot of refactorings that benefited the cause, and changes the
library we use for GraphQL requests since the other one didn't support
this use case.

In addition to printing pull requests in the top level of each
assignee's workload, it also nests linked pull requests under issues as
a sub-list.