observability: searcher: correct query for "unindexed search request errors"
Prior to https://github.com/sourcegraph/sourcegraph/pull/9672 this panel
showed errors "over 5m" (e.g. 1 per second over 5 minutes) which is more difficult
to understand than "every 5m" (e.g. 5 per 5 minutes), so in that PR it was
changed however the query was incorrectly not updated to reflect this.