remove unused support for SOURCEGRAPH_LICENSE_KEY env var
Previously, you could set the license key either in site config (documented and used) or in the `SOURCEGRAPH_LICENSE_KEY` env var (undocumented and unused). This removes the latter method. It was never documented (so is almost certainly not used, although we have no way of knowing for sure), and it offers no clear benefits over putting it in site config.