refactor code for determining a license's features/plan
This refactor has no user-facing behavior changes. Previously, we relied solely on unstructured license tags (such as "starter") to restrict usage. A license with no tags was "Enterprise", effectively allowing all features. That same behavior is preserved. However, this refactor will make it easier in the (very near) future to add multiple plans that have different feature sets (which will be defined in `data.go`).

It also simplifies how feature-checking works (and how to mock that in tests).