refactor license features/plans code (#9849)
* fix path after previous rename

* add tests for FeatureRemoteExtensionsAllowDisallow feature

* rename file to match Go idiomatic file naming

* add read-license utility for printing license information

* remove unused support for SOURCEGRAPH_LICENSE_KEY env var

Previously, you could set the license key either in site config (documented and used) or in the `SOURCEGRAPH_LICENSE_KEY` env var (undocumented and unused). This removes the latter method. It was never documented (so is almost certainly not used, although we have no way of knowing for sure), and it offers no clear benefits over putting it in site config.

* move features to data.go

* remove unused support for license keys containing $FEATURE and no-$FEATURE tags

This tag could be used to enable or disable features that would otherwise be disabled or enabled by a license key. There are no licenses that we've generated that have this kind of tag, according to the following GraphQL API query on Sourcegraph.com:

https://sourcegraph.com/api/console#%7B%22query%22%3A%22query%20%7B%5Cn%20%20dotcom%20%7B%5Cn%20%20%20%20productLicenses%20%7B%5Cn%20%20%20%20%20%20nodes%20%7B%5Cn%20%20%20%20%20%20%20%20info%20%7B%5Cn%20%20%20%20%20%20%20%20%20%20tags%5Cn%20%20%20%20%20%20%20%20%7D%5Cn%20%20%20%20%20%20%7D%5Cn%20%20%20%20%7D%5Cn%20%20%7D%5Cn%7D%5Cn%22%7D

* fix license enforcement message (keep error message prefix)

* remove hardcoded license tags for product names

All products in Stripe now have the licenseTags metadata set for them, so it is not necessary to supply tags based on the product name.

* refactor code for determining a license's features/plan

This refactor has no user-facing behavior changes. Previously, we relied solely on unstructured license tags (such as "starter") to restrict usage. A license with no tags was "Enterprise", effectively allowing all features. That same behavior is preserved. However, this refactor will make it easier in the (very near) future to add multiple plans that have different feature sets (which will be defined in `data.go`).

It also simplifies how feature-checking works (and how to mock that in tests).