include the license key in pings (#9833)
The ping payload now includes the license key, which is a base64-encoded string that describes the premium features enabled for a particular Sourcegraph instance. This will make it easier to track instances that will soon have an expired license or who are using an obsolete license key (e.g., if we upgrade the license key's schema).