debugproxies: Nicer display names for statefulset pods (#9950)
StatefulSet pods have unique hostnames, so we don't need to refer to them by
the IP. This will give them nice and stable names via our site admin debug
proxies page. The nice part of this is we can include sample links to pod -0
in documentation and they should work for customers.