observability: separate search-based code intelligence from regular search metrics
Prior to this commit, "Search" and "Search API" were the only distinction in
the frontend monitoring dashboard. However, I learned that search-based code
intelligence may be the most massive player on Sourcegraph.com's search usage
and currently there is no way to distinguish the two. This makes it so that
the two are distinguished and the breakdown is as follows:

- "Search at a glance" (shown by default)
- "Search-based code intelligence" (hidden by default)
- "Search API usage" (hidden by default)

The actual browser extension change so that the request_type is changed from
/.api/graphql?Search to /.api/graphql?CodeIntelSearch will be rolled out
separately.