add request_name to src_graphql_field_seconds metric
This is needed to support https://github.com/sourcegraph/sourcegraph/pull/9986
I mistakenly thought this metric already included request_name because
of its large cardinality, but I was wrong.

This increases the cardinality of this metric by 1, so we will now have
1802 unique time series. This is already down from 2500 unique time series
thanks to my earlier change in https://github.com/sourcegraph/sourcegraph/issues/9895#issuecomment-615524694

I will additionally further reduce the cardinality of this metric
in order to guarantee this is safe.