observability: add site configuration option to log slow GraphQL and search queries
This is a very low-risk addition that adds two site configuraiton options to log
slow GraphQL and search queries. It is low-risk because it is only running if the
option is set in the site configuration.

We have seen twice in the past that these options would be very useful in debugging,
and right now one customer is asking how they would debug slow search queries and
this would help them.