debugproxies: use IP for addresses if we have it
The hostname for pods is missing the service name in kubernetes, which means
we can't actually resolve the address. So instead we use the IP
again. However, we do store the Hostname if we have it so we can use it for
display purposes. This isn't actually read yet, will be used in the next
commit.