monitoring: do not use a separate go.mod/go.sum file
the monitoring generator used a seperate go.mod since it must have one for dependency pinning
when we transfer it to Docker as the build context (unless we transfer the entire repo as build
context, which is not ideal). However, this seperate go.mod meant that sometimes when editing
the monitoring generator's files in e.g. VS Code the tooling would (unfortunately) modify the
root go.mod/go.sum instead of the one in the monitoring directory. To workaround this, this
change switches us to using the root's and just copying it into the Docker container when
needed.