dev/watchman: ignore hidden files (#10085)
Hidden files are often written to while you are editing (emacs and vim swap
files for example). Currently this would trigger an event in watchman. So we
add an exclude rule for ".*" glob. We take the existing expression and AND it
with the exclude rule. So this change is

  ["allof", ["not", ["match", ".*"]], ${OLD_EXPRESSION}]

Also included in this change is a way to debug which files watchman thinks has
changed. You set the environment variable WATHCMAN_DEBUG to a non-empty value.