graphql: remove discussions
This was deprecated in 3.15 for removal in 3.16. Most of the code still
exists, this is just removing it from the GraphQL backend.

Will need to follow-up with configuration and other parts of the codebase
which have discussions code. Will also confirm that there are no calls to the
graphql endpoint on Sourcegraph.com before merging.