Remove uncommitted transaction from ExecChangesetJob
The code in `ExecChangesetJob` was just there to ensure that a
transaction was opened.

The problem is that it doesn't fully work:

1. If a transaction was already open, the code is a noop. Good.
2. If no transaction was opened, it opens one, but never commits it. Bad.

So instead of using half a solution, we remove the code that can lead to
problems and add comments to ensure the code is called inside a
transaction.