Drop GitHub ReviewRequestedEvents when reviewer has been deleted (#10284)
This is another follow-up to #10264. After fixing one of the symptoms of
deleted reviewers in #10267 I decided to follow @ryanslade's suggestion
and actually "remove" the events.

Now, actually removing the matching entry from the database is not
possible with out current architecture, since it would require us to
persist the GitHub-Node-ID of each event and then, on sync, map the
now-with-deleted-reviewer event to the one in the database and then not
only drop the new event but also delete the one in the DB.

I think we can tackle in the future, when we want to make sure that our
"sync" is actually a full sync (i.e. deletes items in the DB that do not
exist on the code host anymore).

For now, this lets me sleep better at night.