ctags: skip long lines
This is code I wrote for Zoekt's ctags code which prevented us failing to
parse repos with large symbols [1]. The code in our repo was branched off from
the Zoekt code based on the similarties I see, so porting it to ours. This is
in an effort to unify both.

Note: I This shouldn't happen since we introduced CTAGS_PATTERN_LENGTH_LIMIT,
but this brings us closer to having the same ctags code for both zoekt and
symbols service.

[1]: https://gerrit-review.googlesource.com/c/zoekt/+/241732