gitserver: limit internal cloning to 1 Gbps
zoekt-sourcegraph-indexserver currently limits the rate at which a git
archive is fetched to 1 Gbps. This is to prevent us overloading the
network. Before we had the limit we would cause the docker networking
stack to crash. Since zoekt is moving to using our internal clone
endpoint, we want to keep the limiting behaviour. Instead of limiting on
the client side, we limit on the server side. This is done since
introducing limiting in git would require extra tools to be installed.

The comment and limit is the same we have currently for the archive
based approach.