wip search config endpoint for multiple branches
This is a WIP commit illustrating how we will pass down which branches
to index for a repository to Zoekt. It is only documenting which fields
we need to send.

Note: zoekt-sourcegraph-indexserver doesn't yet support this. If you
want you can also update zoekt-sourcegraph-indexserver to support
this. I would suggest that if Branches is non-empty
zoekt-sourcegraph-indexserver would use zoekt-git-index instead of
zoekt-archive-index. This would be a nice way to feature flag.