Make Campaigns teaser pages clearer and easier to follow
This removes a lot of redundant information on the teaser pages
(previously we had multiple mentions of "only available to site admins",
multiple "reach out to us", ...) and makes the getting-started steps
easier to follow.

Instead of having 2 call-to-action buttons and a list of steps that all
link to the same documentation page, I changed it to a single list of
steps that leads to the campaigns documentation's "Getting started"
guide.

Until we can make bigger improvements to the UI that allows us to
incorporate the "Getting started" guide _into the product itself_, I
think it's better to not have redundant guides and provide the user with
a _single_ list of instructions they can follow step-by-step.