Bext: make permissions warning less scary
Context: https://sourcegraph.slack.com/archives/CMT39K56Z/p1589217077358500

A user got concerned when seeing the "grant permissions" notification on a non-GitHub site, thinking that he needed to do it, and uninstalled the browser extension as a result.

This makes the warning more informative, explaining why this is necessary, and on which type of site.