dbtest: Always rollback transaction, even if test failed
I'm not sure why we we never rolled back a transaction, nor committed
it. I suspect the idea was to leave test state available in the test
database. But if the transaction is never committed, that state is never
available anyway.

And with the previous behaviour I can into problems when multiple tests,
such as the ones in `campaigns.TestIntegration`, use the same test
database and one of the tests fails. The failing test then leaves a
transaction dangling, which can block other tests.