gitserver: check clone in progress if not cloned
When we run a git command we first check if there is a clone in progress
for the repo. If there is we return saying the repo is busy
cloning. However, the clone may in fact exist since our cloning logic
clones to a temp directory and renames into place. For example our
reclone script.

This commit instead will use the existing clone if it is there. It does
this by only checking the clone lock if the repo is not in place.