server: remove upgrade migration
We have had this code in Sourcegraph server for a _long_ time but it has
been a noop for most of that time. If we decide to upgrade postgres
versions again we can revert this commit (and do the harder challenge of
supporting upgrades in our cluster environments).

Main motivation for removing it is simplification.