indexed-search: use zoekt.FileMatch.Version for SHA
A while ago we added "IndexedHEADCommit" for finding the sha for a
branch we searched in zoekt. However, the Zoekt API returns the SHA in
the result struct. It will also be the actual sha, and won't suffer from
potential race conditions. We also get the benefit of a smaller struct
for RepositoryRevisions.

This change is additionally required once we have multiple branches
indexed since we will need a way to resolve the commit for more than
just HEAD.

Note: we will now use indexed search if it has multiple
branches. However, we can't turn on multiple branches yet since we do
not yet tell zoekt which branches to search.