Smart search: show all filter value suggestions
Fixes #10770

By default, Monaco filters suggestions based on a suggestion's `label`, and the range targeted by the suggestion. This caused suggestions for, eg., repositories not to show up if the filter value was a regex pattern `^github\.com/`, but the suggestion's label is not `github.com/sourcegraph/sourcegraph`. This is fixed by setting the current filter value as `filterText`, essentially bypassing Monaco's filtering.