search: do not quote rev in dynamic repo filter
We previously quoted rev as if it was a regex. However, it is a
refglob. The only characters we interpret in our parsing layer after the
'@' are '@' and ':'. However, both of those characters are disallowed in
valid git revisions (refs or shas), so we don't need to quote them.

While testing version contexts we had incorrect filters added since we
were exploring the "3.0" branch. It added the rev specifier "3\.0"
instead of just "3.0".