search: include revision in repository search results
If we have a revision for a repository we are searching over, we should
include it in the results. For example if you search for "repo:" in a
version context we should link to the repo at the version in the
context, rather than HEAD.

Additionally this PR allows version contexts to specify a repository
multiple times (to allow searching it at multiple revisions).