search: ParseRepositoryRevisions returns a pattern string
Previously we incorrectly returned an api.RepoName. This was wrong for
two reasons. Firstly it is a regex pattern (ie the input to a repo:
search). Secondly api.RepoName should generally come from the DB (ie
should be an actual repo name, not user input). The removal of type
casting in call sites is more proof that this is the correct type.