gitserver: fetch GitLab and Bitbucket pull refs
We have EnsureRevision logic which will trigger a fetch if a revision is
missing. This will run when browsing a pull request on GitLab. However,
if the pull request is not from a branch in the same repository our
fetch would not fetch the relevant commit. This leads to us constantly
fetching and failing.

In particular we are noticing "git fetch" constantly running for
gitlab.com/gitlab-org/gitlab and in some cases it running very slowly. I
assume this gitlab rate limiting/metering our requests. The metering
leads to an unresponsive Sourcegraph request since we wait upto 60s for
the fetch to happen for a missing revision.