repo-updater: Use per HTTP request timeouts
This commit removes the global 30 minutes sourcing timeout and replaces
it with per HTTP request timeout of 60 seconds which is set in
`NewExternalHTTPClientFactory`.

This per request timeout will apply to all external HTTP requests we do
(i.e. beyond repo-updater), which seems appropriate.

A global timeout for sourcing repos in the code host isn't necessary
since this is a background operation. As it turns out, just listing
repos can be very slow if there are enough repos to be listed.

As seen in the logs in Stackdriver, each page of repos is taking between
900ms and 3 seconds, which quickly adds up to more than the previously
set 30 minutes timeout.