Don't use "blacklist" or "whitelist" (#11386)
We are against racism. "blacklist" and "whitelist" carry the implication that black is bad and white is good.

We should use words that directly describe what we mean.
backlist -> blocklist
whitelist -> allowlist

Other teams are changing this too: https://9to5google.com/2020/06/07/google-chrome-blacklist-blocklist-more-inclusive/

This PR contains the easy changes. There will be some other changes that we need to make in a follow-up PR because they impact user configuration.