Reject campaign update if lacking permission for any of new patches
Previously we simply skipped the patches and didn't publish them if the
user lost permissions for a repo in between creating the patchset and
updating the campaign.

That could lead to an unfixable state where a campaign had a patch and a
changeset for the same repository.

We decided that the best solution is to simply return an error and
reject the update.