gomod: update zoekt (#11432)
We will be using query.RepoBranches to query zoekt once we support
multiple branches. Below is the commit message from zoekt:

RepoBranches is like our RepoSet query atom. Except it allows us to
specify which branches to search in each repository in the set. This is
required for Sourcegraph to efficiently construct queries which search
across different branches. See RFC 150 Version Contexts for more
context.