zoekt: assume all repos are indexed for pre-allocation
As an optimization we set the capacity for the list of indexed
repositories to search. We estimate that capacity via first looping over
the list of repositories to search. However, in multi branch search the
heuristic is not accurate. Additionally the common case is that when the
list of repos to search is large, we are doing an indexed search. So we
can simplify and just always set the cap to the max size.