limit number of ctags Docker containers in dev
By default, the number of ctags Docker containers is equal to the number of CPUs (as returned by Go's `runtime.GOMAXPROCS(0)`). This makes sense for servers. In dev, this is usually too many. My dev machine has 64 CPUs, and this results in 64 Docker containers running, which causes significant network issues on my machine because so many networking interfaces (for Docker container networking) are added.