WIP: cmd/sourcegraph: single-program server entrypoint
The `sourcegraph` program is a single (Go) static binary that runs a Sourcegraph server. It is an experimental way to let us distribute Sourcegraph as a deb/nix/brew package (or as a simple standalone static binary), which some people prefer to a Docker image.

When invoked without arguments, it spawns subprocesses `sourcegraph frontend`, `sourcegraph gitserver`, `sourcegraph repo-updater`, etc., one for each service.

When invoked with an argument such as `sourcegraph frontend`, it runs the `cmd/frontend` program with the appropriate env vars (which are mostly equivalent to what `cmd/server`, the single-Docker-image program, would use).