perf: Do not unnecessarily load patches.diff from DB
Most of the time when we load patches from the database we don't need to
load the `diff` column, since we're mainly interested in foreign keys or
the diff stat columns.

Since the `diff`s can get pretty big, it makes sense to only load them
when necessary.

By not loading them, we should get a speed bump and use less memory for
requests in which we're really only interested in the other columns
(e.g. when we calculate the diff stat).

This should fix #11126.