repo-updater: make syncer & scheduler aware of uncloned gitserver repositories
This is a fix for #11091 which has become a critical blocking issue for a large customer ([details here](https://sourcegraph.slack.com/archives/CJX299FGE/p1592517101301600)).

I chose to stay online several hours late and implement this myself as there is a heavy
sense of urgency here. With that said, the implementation turned out to be trickier than
I thought it would be. It was not 100% clear-cut to me where this logic should be structured
and live, so it may makesense for someone with broader architect design goals in mind here to
evaluate and adjust post-merge. I think what I have done here is not a completely unreasonable
implementation, though (it just doesn't fit as seamlessly as I would like it to).

Prior to this change:

- repo-updater would instruct gitserver to update repositories if they were added,
  removed, or had their metadata changed on the code host.
- gitserver would be responsible for keeping its own repositories up-to-date
  for Git updates by periodically running `git pull`.

However, it was possible for these two to get out of sync. For example from
a completely clean slate when deploying a new Sourcegraph instance:

- repo-updater identifies thousands of new repositories on the code host not yet in the DB store.
- repo-updater instructs gitserver to clone the thousands of new repositories.
- gitserver's in-memory clone queue now has thousands of repositories and begins cloning.
- gitserver is restarted for any reason, its in-memory update queue is now gone.
- gitserver does not have the repositories cloned, it is completely unaware of them.
- repo-updater just naively assumes that gitserver has them cloned and does nothing to correct the situation.

This has particularly bad effects for users & admins:

1. Searches are incredibly slow and timeout because e.g. half of your several thousand repos are not indexed.
   - Why: zoekt can _discover_ that the repositories exist, but [when it attempts to fetch the tar archive it cannot because gitserver says the repository doesn't exist.](https://sourcegraph.com/github.com/sourcegraph/sourcegraph@7fcb1342431195e12e9d6e8242b9eb303540008c/-/blob/cmd/frontend/internal/httpapi/internal.go#L521))
2. The admin UI shows "100,000 repositories enqueued for cloning..." yet it never does anything
   - Why: quite ironically, the status message code *does* compare with Gitserver how many repositories are uncloned - we just do absolutely nothing for that case in the actual code
3. All logs, monitoring, and metrics just show "everything is fine" and "nothing is happening"
   - Why: from the perspective of gitserver, everything _is cloned_ and from the naive perspective of repo-updater _everything is cloned_.
4. There is no good workaround / way to correct this issue
   - Why: You could delete *all* repositories and start over, but cloning takes a long time and you better hope it doesn't happen again (as it did to this customer)